backend/requirements.txt
Django
djangorestframework
# psycopg2-binary
django-bootstrap4
Psycopg2
django-cors-headers
